Bug 781731 - Fixed failed assertion in reftest1; r=doublec

+++ gfx/2d/Blur.h
@@ -59,7 +59,7 @@ public:
 
   AlphaBoxBlur(uint8_t* aData,
                const Rect& aRect,
-               int8_t aStride,
+               int32_t aStride,
                float aSigma);
 
   ~AlphaBoxBlur();
